# integration-tests: Print the output before making the test fail