@blanquicet 2022-04-08
> Your goal is to print more output when tests are runned with `-test.v`, right?
>
No, it also works without using `-test.v`. My goal is to better understand what was the error when the `command.Run()` or `command.Start()` failed. I added an example on the PR description so that it is clearer 🙂.

> If so, maybe you should move this one too: [4d73bfc#diff-aaade3782dd35eb7898f3403d9c0ae1345621a6442ddff5145b0fdae304ed6afR285](https://github.com/kinvolk/inspektor-gadget/commit/4d73bfc25d6537634f7a02ddd63af377d8f15709#diff-aaade3782dd35eb7898f3403d9c0ae1345621a6442ddff5145b0fdae304ed6afR285)
> 
Mmmm the link brings me the commit of this PR, it doesn't show any additional diff 🙁 .